+++ packages/compiler-cli/test/BUILD.bazel
@@ -57,7 +57,7 @@ ts_library(
     testonly = True,
     srcs = [
         "downlevel_decorators_transform_spec.ts",
-        "signal_inputs_metadata_transform_spec.ts",
+        "initializer_api_transforms_spec.ts",
         "signal_queries_metadata_transform_spec.ts",
     ],
     deps = [
